requirements.txt
bert-extractive-summarizer==0.10.1
blis==0.7.9
boto3==1.26.146
botocore==1.29.146
Brotli==1.0.9
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-design-kit==0.0.1
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.6
filelock==3.12.0
Flask==2.2.3
Flask-Compress==1.13
fonttools==4.39.4
fsspec==2023.5.0
huggingface-hub==0.15.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
murmurhash==1.0.9
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.0
pandas==1.3.5
pathy==0.10.1
Pillow==9.5.0
platformdirs==3.5.3
plotly==5.14.0
preshed==3.0.8
protobuf==3.20.3
pycountry==22.3.5
pydantic==1.10.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.29.2
typer==0.7.0
typing_extensions==4.6.3
urllib3==1.26.16
virtualenv==20.23.0
wasabi==1.1.1
Werkzeug==2.2.3
